# backend/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/auth/register": {
      "post": {
        "description": "Crée un utilisateur avec l’e-mail et le mot de passe fournis, puis émet les jetons d’accès et de rafraîchissement.",
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Utilisateur enregistré avec succès",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "La validation a échoué pour les données fournies",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Un compte existe déjà pour cette adresse e-mail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Créer un nouveau compte utilisateur",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "description": "Valide les identifiants, émet les jetons d’accès/de rafraîchissement et positionne le cookie de rafraîchissement.",
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "La validation des identifiants fournis a échoué",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TooManyRequestsResponseDto"
                }
              }
            }
          }
        },
        "summary": "Authentifier un utilisateur par e-mail et mot de passe",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/auth/oauth2/google": {
      "post": {
        "description": "Échange un jeton Google ID contre des jetons applicatifs. Requiert la configuration du service OAuth2.",
        "operationId": "AuthController_loginWithGoogleOAuth",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleOAuthDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Google OAuth n’est pas configuré ou le jeton fourni est invalide",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "Authentifier via Google OAuth",
        "tags": [
          "auth"
        ]
      }
    },
...
  },
  "info": {
    "title": "Area API",
    "description": "API documentation",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "new.user@example.com",
            "description": "Adresse e-mail utilisée pour créer le compte"
          },
          "password": {
            "type": "string",
            "example": "StrongP@ssw0rd!",
            "description": "Mot de passe respectant les exigences (majuscules, minuscules, chiffre, caractère spécial, sans espace)"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginUserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "6f1b88c8-2b8d-4bd4-9cb8-0a6ec9221f2f",
            "description": "Identifiant unique de l’utilisateur authentifié"
          },
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Adresse e-mail de l’utilisateur authentifié"
          },
          "roles": {
            "example": [
              "user"
            ],
            "description": "Rôles attribués à l’utilisateur",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "email",
          "roles"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "Jeton d’accès pour authentifier les requêtes suivantes"
          },
          "expires_in": {
            "type": "number",
            "example": 600,
            "description": "Durée de validité du jeton d’accès en secondes"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer",
            "description": "Type de jeton émis",
            "default": "Bearer"
          },
          "user": {
            "$ref": "#/components/schemas/LoginUserDto"
          },
          "message": {
            "type": "string",
            "example": "Utilisateur créé avec succès",
            "description": "Message de confirmation renvoyé après une inscription réussie"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "token_type",
          "user",
          "message"
        ]
      },
      "ValidationErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Requête invalide",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "example": [
              "le nom ne doit pas être vide",
              "l’identifiant d’action doit être un entier"
            ],
            "description": "Messages de validation détaillés",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "Requête invalide",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "type": "object",
            "example": "Action invalide",
            "description": "Explication lisible de l’erreur. Peut être une chaîne ou un tableau de chaînes en cas d’erreurs de validation."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "user@example.com",
            "description": "Adresse e-mail utilisée pour la connexion"
          },
          "password": {
            "type": "string",
            "example": "StrongP@ssw0rd!",
            "description": "Mot de passe de l’utilisateur"
          },
          "rememberMe": {
            "type": "boolean",
            "example": true,
            "description": "Si vrai, un jeton de rafraîchissement de plus longue durée est émis (si le client le supporte)."
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "example": "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9...",
            "description": "Jeton d’accès pour authentifier les requêtes suivantes"
          },
          "expires_in": {
            "type": "number",
            "example": 600,
            "description": "Durée de validité du jeton d’accès en secondes"
          },
          "token_type": {
            "type": "string",
            "example": "Bearer",
            "description": "Type de jeton émis",
            "default": "Bearer"
          },
          "user": {
            "$ref": "#/components/schemas/LoginUserDto"
          }
        },
        "required": [
          "access_token",
          "expires_in",
          "token_type",
          "user"
        ]
      },
      "UnauthorizedResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 401
          },
          "error": {
            "type": "string",
            "example": "Non autorisé",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "type": "object",
            "example": "Identifiants invalides",
            "description": "Explication lisible de l’erreur. Peut être une chaîne ou un tableau de chaînes en cas d’erreurs de validation."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
      "TooManyRequestsResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 429
          },
          "error": {
            "type": "string",
            "example": "Trop de requêtes",
            "description": "Brève description de la classe d’erreur HTTP"
          },
          "message": {
            "type": "object",
            "example": "Trop de tentatives de connexion. Réessayez plus tard.",
            "description": "Explication lisible de l’erreur. Peut être une chaîne ou un tableau de chaînes en cas d’erreurs de validation."
          }
        },
        "required": [
          "statusCode",
          "error",
          "message"
        ]
      },
      "GoogleOAuthDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "example": "ya29.a0AfH6SMC...",
            "description": "Google ID token (JWT) obtained from the client-side flow"
          }
        },
        "required": [
          "token"
        ]
      },
...
    }
  }
}